Add an entry about daemonization on Linux
More than 90% of the people who come asking about this are likely to be using a Linux system which doesn't have systemd.  Apparently some such individuals can't do any research beyond Syncthing documentation and the issue tracker, so add an entry to the FAQ to give them some reasonable pointers on what to look at to figure out how to do this on their own without bugging the developers.