specs/relay-v1: Clarify relay spec (fixes #488) (#489)
Change the wording for the SessionInvitation message to indicate that the port is not optional. The syncthing reference implementation assumes that the port will be provided, and the reference implementation servers send the port every time.

The client sends a JoinSessionRequest to the relay once it enters session mode. The SessionInvitation is sent only by the relay to the client.