Draft REST API for pending devices and folders.
Basically mirror the information that currently lives within the GET
/rest/system/config endpoint, using two new endpoints under
/rest/system.

Note that for pendingFolders, the structure is inverted: In the
config, they live underneath the "devices" tree, here the offering
devices are listed underneath the pendingFolders entry with the same
information.

Include some example content (totally made up for now as there is no
implementation yet).