rest: Draft API for candidate devices.
Add an endpoint GET /rest/system/candidates to serve the collected
information from ClusterConfig messages about devices with which we
share a folder only indirectly.

Needs either a folder or device ID as parameter, listing the candidate
devices for the folder, or the candidate folders for an already known
device.