users/autostart: Add options to run on battery in Task Scheduler
Add new optional settings for the Conditions tab in Task Scheduler.
Despite the parent-child relationship in the GUI, these two options are
in fact separate, i.e. disabling the parent does not disable the child.
Therefore, "Stop if the computer switches to battery power" needs to be
unchecked first, and "Start the task only if the computer is on AC power"
later.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>