users/autostart: Add option not to run on Wi-Fi in Task Scheduler
It is possible to set Task Scheduler to start a task only when a
specific network connection is available. This way, we can make
Syncthing start only on wired connections, and not on others, e.g.
Wi-Fi. This can be done, because Windows labels wired connections
"Network", while wireless connections all have their own SSIDs.

This still has some limitations, especially when there are multiple
wired network adapters installed, but should be enough for typical
desktop or laptop computers.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>