users/autostart: Remove unneeded recommendations from Settings in Task Scheduler
Both "Allow task to be run on demand" and "Do not start a new instance"
are checked by default, and there is no need to touch them, unless the
user specifically wants to. Therefore, it is unnecessary to specifically
recommend having them checked. Moreover, Syncthing will still run in
background on system start anyway, regardless if the user decides to
fiddle with these two options or not.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>