users/autostart: Improve configuration for Task Scheduler in Windows (#510)
* users/autostart: Recommend not to save passwords in Task Scheduler

Saving user passwords is only required to access non-local resources,
such as network drives, etc. This may present a security a risk, and
also should not be needed in a standard Syncthing installation, where
the folders are located on local drives.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* users/autostart: Add optional 'Start in' path to Task Scheduler

Adding a specific directory to the 'Start in' field in Task Scheduler
allows to use folder paths relative to this folder in Syncthing. If
unset, Windows defaults to %SystemRoot%\System32, which may lead to
unexpected consequences for the unaware users.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* users/autostart: Remove command line arguments from Task Scheduler

With "Run whether user is logged on or not", Task Scheduler starts
processes in background without any visible GUI. Therefore, the
arguments -no-console and -no-browser are superfluous, since neither
a command line window nor a browser window opens in this state.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* users/autostart: Add options to run on battery in Task Scheduler

Add new optional settings for the Conditions tab in Task Scheduler.
Despite the parent-child relationship in the GUI, these two options are
in fact separate, i.e. disabling the parent does not disable the child.
Therefore, "Stop if the computer switches to battery power" needs to be
unchecked first, and "Start the task only if the computer is on AC power"
later.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* users/autostart: Add option not to run on Wi-Fi in Task Scheduler

It is possible to set Task Scheduler to start a task only when a
specific network connection is available. This way, we can make
Syncthing start only on wired connections, and not on others, e.g.
Wi-Fi. This can be done, because Windows labels wired connections
"Network", while wireless connections all have their own SSIDs.

This still has some limitations, especially when there are multiple
wired network adapters installed, but should be enough for typical
desktop or laptop computers.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* users/autostart: Remove unneeded recommendations from Settings in Task Scheduler

Both "Allow task to be run on demand" and "Do not start a new instance"
are checked by default, and there is no need to touch them, unless the
user specifically wants to. Therefore, it is unnecessary to specifically
recommend having them checked. Moreover, Syncthing will still run in
background on system start anyway, regardless if the user decides to
fiddle with these two options or not.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* users/autostart: Add punctuation to instructions for Task Scheduler

Add full stops after each sentence in the Task Scheduler instructions.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* users/autostart: Fix capital letters in instructions for Task Scheduler

Each sentence should begin with a capital letter.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>