users/autostart: Rework Task Scheduler configuration
Task Scheduler is capable of running Syncthing in multiple ways. It is
a good idea to show a way to run it "at log on", and also "on startup".
This way, the user can choose whether they want to start and use
Syncthing while being logged on, or whether they want to have Syncthing
running in background all the time, started right after Windows boots,
and even when the computer is being used by other people.

Furthermore, the "run at log on" option is a good alternative to using
the Startup folder in the Start Menu, especially since Microsoft have
made access to the folder difficult in Windows 10. The folder is
currently not displayed in the Start Menu any more, and can be accessed
only by navigating manually to its actual location inside %APPDATA%,
which itself is a hidden folder, invisible to regular Windows users.

In addition, due to the complexity of Task Scheduler, add screenshots
to clearly show each configuration option on the screen, so that even
beginners should be able to follow the instructions easily.

The actual modifications are as follows:

- Add a short explanation on Task Scheduler itself with a link to the
  official documentation by Microsoft.
- Add simple instructions on how to start Task Scheduler.
- Add a new option describing how to run Syncthing "at system startup"
  in addition to the current "run at log on".
- Add additional explanations to each of the two methods.
- Move optional settings to a new section called "Additional
  configuration in Task Scheduler".
- Add new additional settings to optionally run Syncthing on battery
  power, and using a specific network connection.
- Add screenshots to each step of the instructions.
- Fix the language to match the terminology used by Microsoft (e.g. log
   in -> log on).
- Add cosmetic fixes (indentation, code marking, grammar).

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>