dev/issues: Update labels to match GitHub issues (fixes #559) (#564)
Update the labels with their description to match the actual labels used
currently in the issues in the main Syncthing repository. Replace the
obsolete and unused "unreleased" with "skip-changelog", and mark the
"pr-merged" label as legacy kept for historical purposes. For all the
labels, use clear and easy to understand short descriptions.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>