users/faq: Explain case-sensitivity principles.
State the underlying assumptions and conditions which could lead to a
case conflict.  Advise how to correctly remedy such conflicts and give
some rationale why it's better than pre-v1.9.0 behavior.

Include a warning that the `caseSensitiveFS` advanced option does not
change the basic principles and should thus not be used.  Link to the
corresponding option description.