users/syncing: Explain case-sensitivity principles (fixes #566) (#568)
State the underlying assumptions and conditions which could lead to a
case conflict.  Advise how to correctly remedy such conflicts and give
some rationale why it's better than pre-v1.9.0 behavior.

Include simpler example with only two devices, but keep three-device
case for the fact that the "winning" side is essentially
non-determenistic.

Link to the new section from the `caseSensitiveFS` advanced option
page, within the warning that it does not change the basic principles
and should thus not be used.