users/contrib: Move Syncthing Lite to unmaintained projects (fixes #586)
Syncthing Lite has not been updated for a long time, and its repository
has also been archived. Unless someone decides to step up and continue
working on it, it should be moved to the unmaintained projects list for
now, especially since it does not seem to be functional in its current
state.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>