users/faq: Add a warning about mixing nested folders with Introducer
Add a warning to the FAQ about using the Introducer feature together
with nested folders, as that can lead to Syncthing automatically sharing
both the nested folder itself and its parent between the same devices.

Even though Syncthing can deal with such situations and still properly
synchronise the files, it leads to an unnecessary overhead and creates
temporary files, which hang around for 24 hours (by default) for no
purpose.

Ref: https://github.com/syncthing/syncthing/issues/7338

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>