users/versioning: Improve external versioning example script for Windows
Enable UTF-8 support, as otherwise the script will fail with non-native
code pages, e.g. when using non-Latin characters in the English version
of Windows, etc. Use quotes with the "set" command or the script will be
unable to deal with funky characters like "&" in file and folder names.
Use "md" instead of "mkdir", because it is shorter. Lastly, use "rem"
instead of "::" for comments, as this is the only way to comment that
is recommended by Microsoft (and "::" is just a hack in itself).

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>