users/versioning: External Versioning path need not be absolute
The external versioning path does not necessarily need to be absolute.
For instance, it is possible to a) use a program included in the system
PATH, such as cmd.exe or powershell.exe in Windows, to embed the whole
command, or b) when not specified, the path is relative to the folder,
where the Syncthing binary has been executed.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>