autostart: Clarify that installing systemd units might be unneeded.
In many cases the systemd unit files are already installed, yet users
try to follow these instructions because they haven't read the
paragraph just above or didn't realize it applies to them.  So be more
explicit in the step-by-step instructions that copying unit files may
be unnecessary and one should first try without.