Explain the different CLI modes based on sub-commands.
Add a section about the general concept and include "decrypt" and
"cli" in the syncthing command synopsis.  Top-level entry points into
the "cli" hierarchy are documented in a separate included snippet.

Hopefully something like that can be auto-generated in the distant
future to make sure an up-to-date listing is included in the
documentation.