_ext: Add custom Sphinx extension for configuration cross-refs.
Implement a Sphinx extension with its own domain "stconf" and the
contained custom directive .. stconf:option: and role :stconf:opt:
They work quite similar to the standard Sphinx mechanisms for
documenting functions, classes, etc. including the usual
cross-referencing syntax.  The directive removes the first part
separated by a dot and handles it as a "section" to disambiguate
option names such as "id".

Add the plugin to the Sphinx config file.