Redirect to version-specific subdirectory from query parameter.
Include a custom JavaScript snippet in the built HTML version.  It
checks the ?version=... query parameter and heuristically finds a best
match among the deployed versions list in JSON format.

If a match is found, the browser URL is reconstructed to prepend the
matched version string as a path component.  If the previous first
path component matches a version tag from the list, that is removed
first.  The query parameters get discarded, and the fragment
identifier (#anchor) appended again.  This changed URL is used for a
silent redirect.

The version matching heuristic accepts any version string starting
with a complete known version tag, checking numerically higher ones
first.  For example if v0.12.10 matches first, v0.12.1 is not even
considered anymore.  Incomplete matches are accepted as long as they
only differ after the last dot: v0.13.10 applies to any
v0.13.* (higher or lower) as long as there is no better, direct match.

Any other string in the query parameter leads to the same path with
any known version string in the top-level component removed.  That
should be the default, fall-back version.