Add a header note allowing to pick a specific version.
Install a handler for the DOMContentLoaded event from the existing
version_redirect.js script, which creates and inserts the actual note
box and selector at the top of the "document" div.  This allows to
keep all the historically generated HTML files untouched, except for
including the script.

Using the browser's back button should allow to go back to the
previously displayed version.  It doesn't make sense for URL-parameter
based redirects, though, so keep it off for that case.

Rename and rework splitVersionPath() function.