Add version picker box at the top (#721)
* Add a header note allowing to pick a specific version.

Install a handler for the DOMContentLoaded event from the existing
version_redirect.js script, which creates and inserts the actual note
box and selector at the top of the "document" div.  This allows to
keep all the historically generated HTML files untouched, except for
including the script.

* Fix JS warnings about not well-formed XML.

Using the $.getJSON() method apparently fails to apply the correct
MIME type, trying to parse the script as HTML / XML and giving an
error message.  Especially Firefox seems to be affected by this.
Overriding the MIME type explicitly before each request silences the
error message.  Taken from:
https://stackoverflow.com/questions/2618959/not-well-formed-warning-when-loading-client-side-json-in-firefox-via-jquery-aj

* Keep history when using version picker.

Using the browser's back button should allow to go back to the
previously displayed version.  It doesn't make sense for URL-parameter
based redirects, though, so keep it off for that case.