Match description field names to definition (#806)
In the description, the field names should match the names defined in
the proto definition for better discoverability and less mental load.

When trying to reading and understand the protocol it should be possible
to look for the description of a field by searching the name of such a
field in the document. However, at the moment, if a reader would like to
find out more about the field `encryption_password_token` they should
scroll down until they find `enc pw token` and a little extra effort is
needed to ensure that this the same as `encryption_password_token`. By
using the same name a reader can search and find the field in the
document without extra scrolling.