users: Correctly describe how to use Unix sockets for GUI (#827)
[the docs for
`--gui-address`](https://docs.syncthing.net/users/syncthing.html#cmdoption-gui-address)
incorrectly state that if a file path is given, a unix socket at that
location will be used.

syncthing supports serving the gui and api over a unix socket. this can
be configured via an cli or environment override. this commit corrects
the documentation on this to match the way that syncthing behaves.

the gui-address cli option needs an url like "unix://<path>" to listen
on unix sockets. if an (absolute) path is specified, the cli override is
ignored and the gui.address from the config file is used. I think this
was done because syncthing supports https (http over tls) over unix, and
there needed to be a way to enable that. in the config file there is an
attribute `tls` on the `gui` element. on the cli this is encoded in the
url scheme: "unixs" and "https" for tls, "unix" and "http" for plain
listeners.

---------

Co-authored-by: digital <didev@dinid.net>
Co-authored-by: Jakob Borg <jakob@kastelo.net>