Fix style and recommend OpenSSH component under Windows.
Fix inconsistent capitalization and markup.  Show TLS enabled in the
config example, since that is what we're recommending immediately
after.  Point to an official Microsoft document for getting started
with OpenSSH instead of WSL or GNU On Windows, as the former is now
widely available and much easier to install.