users/config: be more precise about trafficClass value
Although described, it was not obvious to the author
what exact value needs to be specified. To improve this,
it was explicitly stated that it's about the respective
IPv4 and IPv6 header fields.

Additionally, users most likely want to specify a DSCP
value. To help in this case the required bit shift was
mentioned.