users/config: be more precise about trafficClass value (#862)
Although described, it was not obvious to me what exact value needs to
be specified. To improve this, it was explicitly stated that it's about
the respective [IPv4](https://en.wikipedia.org/wiki/Type_of_service) and
[IPv6](https://en.wikipedia.org/wiki/IPv6_packet#Fixed_header) header
fields.

Additionally, users most likely want to specify a
[DSCP](https://en.wikipedia.org/wiki/Differentiated_services) value. To
help in this case the required bit shift was mentioned.