feat(faq): add information on setting network to private in Windows (#914)
The issue of Windows devices being unable to connect directly with each
other comes up often on the forum. One of them main reasons for the
problem is Windows setting the default network profile to "Public"
instead of "Private", which prevents direct connections.

Instead of repeating the same instructions on how to change the network
profile to "Private" over and over, add a new entry to the FAQ, which we
can then use as a reference when needed.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>