chore(specs): Update BEP regarding multiple cluster config messages
Syncthing is already sending multiple cluster configs for a long time -
reflect that in the spec. I added the SHOULD part about closing the
connection if multiple CCs aren't supported with the intention to keep
the spec change backwards compatible and have reasonable behaviour,
which this imo provides.

I noticed we don't specify what the compatibility expectations can be on
the BEP - I just assumed a broad, rather generous variant, i.e. that the
use-case outlined in the intro continuous to function between to clients
implementing the spec before and after a change. Of course in reality
only the connection to syncthing's compatibility guarantee really
matters, i.e. that two syncthing versions before and after keep syncing
fine, as there's no other significantly adopted BEP implementations - I
just still have a soft spot for the mere possibility of independent
implementations :)