feat(autostart): add note about allowing access in Windows Firewall
Currently, we always recommend to "Run whether user is logged on or not"
when using the Task Scheduler to start Syncthing automatically. The
problem is that this makes Syncthing run completely hidden in background
which prevents Windows Firewall from prompting the user to allow
Syncthing to access the network. Therefore, add a note explaining that
the user should run Syncthing interactively at least once, which will
allow the Windows Firewall pop-up to be displayed on the screen.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>