fix(autostart): remove --no-console from Windows instructions
With the advent of Windows Terminal and it being the default shell both
in newer releases of Windows 10, and in Windows 11, the `--no-console`
command line switch no longer works. For this reason, remove it from the
commands listed in the autostart instructions. When it comes to the Task
Scheduler part, it has never actually been needed anyway, and in the
Startup folder instructions, it can be worked around by using the native
`conhost.exe --headless` method.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>
Co-authored-by: Rafael Ledo <rafaeloliveiraledo@gmail.com>
Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>