fix(specs): fix relay formatting errors (#970)
First error (introduced in #961) fixed is where the "The relay
immediately terminates the connection after sending this." is not
included in the enumeration above. That statement only applies in the
case of RelayFull.

Second error is that the response codes right after the "Protocol
defined responses" comment are actually considered a part of the comment
because the indentation level is 1 higher, and therefore won't be
displayed.

According to
https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#comments:

> Arbitrary indented text may be used on the lines following the
explicit markup start:
> ```
> .. This is a comment
> ..
>    _so: is this!
> ```

---------

Signed-off-by: Marcus B Spencer <marcus@marcusspencer.us>